Address code review feedback
- Split normalize.rs (538 lines) into modules:
  - Created validate.rs with validation functions (validate_formula_semantics,
    validate_invalid_not_in_or, validate_positive_terms)
  - Moved tests to tests/normalize_tests.rs
- Fixed LegacyClause::Constraint to return not-implemented diagnostic instead
  of silently skipping constraints
- Verified MatchFormula::Decorated preserves as_name and fix decorations
- Fixed normalization_tests.rs to panic on file I/O errors instead of using
  semantically wrong ESempaiYamlParse diagnostic code
- Updated feature file scenario title to reflect actual behavior
- Removed trailing commas from execplan bullet points
- Merged duplicate # Errors sections in engine.rs compile_yaml doc comment
