+++ crates/sempai/src/lib.rs
@@ -39,6 +39,7 @@ mod engine;
 
 mod mode_validation;
 pub(crate) mod normalize;
+mod validate;
 
 // Re-export all stable types from sempai_core.
 pub use sempai_core::{
