+++ crates/sempai/src/normalize.rs
@@ -6,6 +6,8 @@ use sempai_yaml::{
     RulePrincipal, SearchQueryPrincipal,
 };
 
+use crate::validate::validate_formula_semantics;
+
 /// A normalized search rule ready for query plan construction.
 #[derive(Debug, Clone)]
 pub struct NormalizedSearchRule {
@@ -170,11 +172,11 @@ fn normalize_legacy_clause(
         }
         LegacyClause::Constraint(_) => {
             // Constraints (metavariable-pattern, metavariable-regex, etc.) are
-            // not yet implemented. We return None to skip them entirely,
-            // rather than creating an empty And which would cause semantic
-            // validation to fail. In a full implementation, we'd parse them
-            // into WhereClause variants and preserve them.
-            Ok(None)
+            // not yet implemented. Return a not-implemented diagnostic
+            // instead of silently skipping them.
+            Err(DiagnosticReport::not_implemented(
+                "Legacy constraint normalization (metavariable-pattern, metavariable-regex, etc.)",
+            ))
         }
     }
 }
@@ -239,94 +241,6 @@ fn normalize_v2_formula(formula: &MatchFormula) -> Result<DecoratedFormula, Diag
     }
 }
 
-/// Validates semantic constraints on the normalized formula.
-///
-/// This checks:
-/// - `InvalidNotInOr`: Disjunctions cannot have negated branches
-/// - `MissingPositiveTermInAnd`: Conjunctions must have at least one positive term
-fn validate_formula_semantics(formula: &Formula) -> Result<(), DiagnosticReport> {
-    validate_invalid_not_in_or(formula)?;
-    validate_positive_terms(formula)?;
-    Ok(())
-}
-
-/// Validates that disjunctions do not contain negated branches.
-fn validate_invalid_not_in_or(formula: &Formula) -> Result<(), DiagnosticReport> {
-    match formula {
-        Formula::Or(children) => {
-            for child in children {
-                if matches!(child.formula, Formula::Not(_)) {
-                    return Err(DiagnosticReport::single_error(
-                        DiagnosticCode::ESempaiInvalidNotInOr,
-                        "negation is not allowed inside 'pattern-either' or 'any'".to_owned(),
-                        None,
-                        vec![
-                            "Move the negation outside the disjunction, or restructure the query"
-                                .to_owned(),
-                        ],
-                    ));
-                }
-                // Recursively check nested formulas
-                validate_invalid_not_in_or(&child.formula)?;
-            }
-            Ok(())
-        }
-        Formula::And(children) => {
-            for child in children {
-                validate_invalid_not_in_or(&child.formula)?;
-            }
-            Ok(())
-        }
-        Formula::Not(inner) | Formula::Inside(inner) | Formula::Anywhere(inner) => {
-            validate_invalid_not_in_or(&inner.formula)
-        }
-        Formula::Atom(_) => Ok(()),
-    }
-}
-
-/// Validates that conjunctions have at least one positive term.
-fn validate_positive_terms(formula: &Formula) -> Result<(), DiagnosticReport> {
-    match formula {
-        Formula::And(children) => {
-            // Empty conjunctions are allowed (they represent no-op/placeholder formulas)
-            if children.is_empty() {
-                return Ok(());
-            }
-
-            // Check if there's at least one positive term
-            let has_positive = children.iter().any(DecoratedFormula::is_positive_term);
-
-            if !has_positive {
-                return Err(DiagnosticReport::single_error(
-                    DiagnosticCode::ESempaiMissingPositiveTermInAnd,
-                    "conjunction must contain at least one positive match term".to_owned(),
-                    None,
-                    vec![
-                        "Add a 'pattern' or 'regex' term to the conjunction".to_owned(),
-                        "Note: 'inside', 'anywhere', and 'not' are constraints, not match producers".to_owned(),
-                    ],
-                ));
-            }
-
-            // Recursively validate children
-            for child in children {
-                validate_positive_terms(&child.formula)?;
-            }
-            Ok(())
-        }
-        Formula::Or(children) => {
-            for child in children {
-                validate_positive_terms(&child.formula)?;
-            }
-            Ok(())
-        }
-        Formula::Not(inner) | Formula::Inside(inner) | Formula::Anywhere(inner) => {
-            validate_positive_terms(&inner.formula)
-        }
-        Formula::Atom(_) => Ok(()),
-    }
-}
-
 /// Normalizes language strings from the rule into Language enums.
 fn normalize_languages(languages: &[String]) -> Result<Vec<Language>, DiagnosticReport> {
     languages
@@ -342,197 +256,4 @@ fn normalize_languages(languages: &[String]) -> Result<Vec<Language>, Diagnostic
             })
         })
         .collect()
-}
-
-#[cfg(test)]
-#[expect(clippy::unwrap_used, reason = "tests use unwrap for brevity")]
-#[expect(clippy::indexing_slicing, reason = "tests panic on out-of-bounds")]
-mod tests {
-    use super::*;
-
-    fn parse_and_normalize(yaml: &str) -> Result<Vec<NormalizedSearchRule>, DiagnosticReport> {
-        let file =
-            sempai_yaml::parse_rule_file(yaml, None).expect("test fixture YAML should parse");
-        normalize_rule_file(&file)
-    }
-
-    fn expect_diagnostic(yaml: &str, expected_code: DiagnosticCode) {
-        let err = parse_and_normalize(yaml).expect_err("should fail");
-        let first = err
-            .diagnostics()
-            .first()
-            .expect("should have at least one diagnostic");
-        assert_eq!(first.code(), expected_code);
-    }
-
-    #[test]
-    fn normalize_simple_pattern_legacy() {
-        let yaml = concat!(
-            "rules:\n",
-            "  - id: test.rule\n",
-            "    message: test\n",
-            "    languages: [rust]\n",
-            "    severity: ERROR\n",
-            "    pattern: fn $F($X)\n",
-        );
-        let result = parse_and_normalize(yaml).unwrap();
-
-        assert_eq!(result.len(), 1);
-        assert_eq!(result[0].rule_id, "test.rule");
-        assert_eq!(result[0].language, Language::Rust);
-        assert!(matches!(result[0].formula, Formula::Atom(Atom::Pattern(_))));
-    }
-
-    #[test]
-    fn normalize_simple_pattern_v2() {
-        let yaml = concat!(
-            "rules:\n",
-            "  - id: test.rule\n",
-            "    message: test\n",
-            "    languages: [rust]\n",
-            "    severity: ERROR\n",
-            "    match: fn $F($X)\n",
-        );
-        let result = parse_and_normalize(yaml).unwrap();
-
-        assert_eq!(result.len(), 1);
-        assert_eq!(result[0].rule_id, "test.rule");
-        assert_eq!(result[0].language, Language::Rust);
-        assert!(matches!(result[0].formula, Formula::Atom(Atom::Pattern(_))));
-    }
-
-    #[test]
-    fn legacy_and_v2_patterns_normalize_equivalently() {
-        let legacy_yaml = concat!(
-            "rules:\n",
-            "  - id: test.rule\n",
-            "    message: test\n",
-            "    languages: [rust]\n",
-            "    severity: ERROR\n",
-            "    pattern: fn $F($X)\n",
-        );
-        let v2_yaml = concat!(
-            "rules:\n",
-            "  - id: test.rule\n",
-            "    message: test\n",
-            "    languages: [rust]\n",
-            "    severity: ERROR\n",
-            "    match: fn $F($X)\n",
-        );
-
-        let legacy_result = parse_and_normalize(legacy_yaml).unwrap();
-        let v2_result = parse_and_normalize(v2_yaml).unwrap();
-
-        assert_eq!(legacy_result[0].formula, v2_result[0].formula);
-    }
-
-    #[test]
-    fn invalid_not_in_or_detected_legacy() {
-        expect_diagnostic(
-            concat!(
-                "rules:\n",
-                "  - id: test.rule\n",
-                "    message: test\n",
-                "    languages: [rust]\n",
-                "    severity: ERROR\n",
-                "    pattern-either:\n",
-                "      - pattern-not: fn $F($X)\n",
-                "      - pattern: fn $G($Y)\n",
-            ),
-            DiagnosticCode::ESempaiInvalidNotInOr,
-        );
-    }
-
-    #[test]
-    fn invalid_not_in_or_detected_v2() {
-        expect_diagnostic(
-            concat!(
-                "rules:\n",
-                "  - id: test.rule\n",
-                "    message: test\n",
-                "    languages: [rust]\n",
-                "    severity: ERROR\n",
-                "    match:\n",
-                "      any:\n",
-                "        - not:\n",
-                "            pattern: fn $F($X)\n",
-                "        - pattern: fn $G($Y)\n",
-            ),
-            DiagnosticCode::ESempaiInvalidNotInOr,
-        );
-    }
-
-    #[test]
-    fn missing_positive_term_detected_legacy() {
-        expect_diagnostic(
-            concat!(
-                "rules:\n",
-                "  - id: test.rule\n",
-                "    message: test\n",
-                "    languages: [rust]\n",
-                "    severity: ERROR\n",
-                "    patterns:\n",
-                "      - pattern-not: fn $F($X)\n",
-                "      - pattern-inside: impl $T\n",
-            ),
-            DiagnosticCode::ESempaiMissingPositiveTermInAnd,
-        );
-    }
-
-    #[test]
-    fn missing_positive_term_detected_v2() {
-        expect_diagnostic(
-            concat!(
-                "rules:\n",
-                "  - id: test.rule\n",
-                "    message: test\n",
-                "    languages: [rust]\n",
-                "    severity: ERROR\n",
-                "    match:\n",
-                "      all:\n",
-                "        - not:\n",
-                "            pattern: fn $F($X)\n",
-                "        - inside:\n",
-                "            pattern: impl $T\n",
-            ),
-            DiagnosticCode::ESempaiMissingPositiveTermInAnd,
-        );
-    }
-
-    #[test]
-    fn unsupported_mode_returns_error() {
-        expect_diagnostic(
-            concat!(
-                "rules:\n",
-                "  - id: test.rule\n",
-                "    mode: taint\n",
-                "    message: test\n",
-                "    languages: [rust]\n",
-                "    severity: ERROR\n",
-                "    taint:\n",
-                "      sources: []\n",
-                "      sinks: []\n",
-            ),
-            DiagnosticCode::ESempaiUnsupportedMode,
-        );
-    }
-
-    #[test]
-    fn multi_language_rule_expands_to_multiple_rules() {
-        let yaml = concat!(
-            "rules:\n",
-            "  - id: test.rule\n",
-            "    message: test\n",
-            "    languages: [rust, python]\n",
-            "    severity: ERROR\n",
-            "    pattern: fn $F($X)\n",
-        );
-        let result = parse_and_normalize(yaml).unwrap();
-
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0].language, Language::Rust);
-        assert_eq!(result[1].language, Language::Python);
-        // Both should have the same formula
-        assert_eq!(result[0].formula, result[1].formula);
-    }
 }
