+++ crates/sempai/src/tests/mod.rs
@@ -2,6 +2,7 @@
 
 mod engine_tests;
 mod normalization_tests;
+mod normalize_tests;
 mod reexport_tests;
 
 mod behaviour;
