+++ crates/sempai/src/tests/normalization_tests.rs
@@ -1,6 +1,7 @@
 //! Tests for legacy/v2 formula normalization into canonical `Formula` model.
 
 #![expect(clippy::indexing_slicing, reason = "tests panic on out-of-bounds")]
+#![expect(clippy::panic_in_result_fn, reason = "test infrastructure may panic on fixture I/O errors")]
 
 use std::fs;
 use std::path::PathBuf;
@@ -20,17 +21,8 @@ fn normalize_fixture(
     filename: &str,
 ) -> Result<Vec<crate::normalize::NormalizedSearchRule>, DiagnosticReport> {
     let path = fixtures_dir().join(filename);
-    let yaml = match fs::read_to_string(&path) {
-        Ok(content) => content,
-        Err(e) => {
-            return Err(DiagnosticReport::single_error(
-                sempai_core::DiagnosticCode::ESempaiYamlParse,
-                format!("failed to read fixture {filename}: {e}"),
-                None,
-                vec![],
-            ));
-        }
-    };
+    let yaml = fs::read_to_string(&path)
+        .unwrap_or_else(|e| panic!("failed to read fixture {filename}: {e}"));
     let uri = path.to_str().map(String::from);
     let file = parse_rule_file(&yaml, uri.as_deref())?;
     normalize_rule_file(&file)
